+++ mmtview/src/loadtest/LoadTestPanel.tsx
@@ -6,7 +6,6 @@ import { FlowchartView } from '../flowchart';
 import { FileContext } from '../fileContext';
 import PanelRunHeader, { HeaderAction } from '../components/PanelRunHeader';
 import PanelEditHeader from '../components/PanelEditHeader';
-import { HideWhenYamlError } from '../api/YamlErrorWarning';
 
 interface LoadTestPanelProps {
   content: string;
@@ -57,13 +56,11 @@ const LoadTestPanel: React.FC<LoadTestPanelProps> = ({ content, setContent }) =>
                         onClick={() => setPage('flow')}
                         disabled={Boolean(flowchartState?.noItems)}
                       />
-                      <HideWhenYamlError>
                         <HeaderAction
                           icon="edit"
                           label="Edit Load Test"
                           onClick={() => setPage('edit')}
                         />
-                      </HideWhenYamlError>
                     </>
                   }
                 />